Split network controller into a cpp file
